refactor customize query accordion body tables use semantic html table elements (#2633)
* seed the new table component

* [pre-commit.ci] auto fixes from pre-commit hooks

* cleanup some unneeded files

* lint

---------
