+++ containers/tefca-viewer/src/app/query/components/customizeQuery/CustomizeQueryAccordionBody.tsx
@@ -1,6 +1,7 @@
 import { Icon } from "@trussworks/react-uswds";
 import styles from "./customizeQuery.module.css";
 import { GroupedValueSet } from "./customizeQueryUtils";
+import Table from "../../designSystem/Table";
 
 type CustomizeQueryAccordionBodyProps = {
   group: GroupedValueSet;
@@ -22,39 +23,37 @@ const CustomizeQueryAccordionBody: React.FC<
   CustomizeQueryAccordionBodyProps
 > = ({ group, toggleInclude, groupIndex }) => {
   return (
-    <div className={`${styles.customizeQueryAccordion__body}`}>
-      <div className={`${styles.customizeQueryGridContainer}`}>
-        <div className={`${styles.customizeQueryGridHeader} margin-top-10`}>
-          <div className={`${styles.accordionTableHeader}`}>Include</div>
-          <div className={`${styles.accordionTableHeader}`}>Code</div>
-          <div className={`${styles.accordionTableHeader}`}>Display</div>
-        </div>
-        <div className="display-flex flex-column">
-          {group.items.map((item, index) => (
-            <div className={`${styles.customizeQueryGridRow}`} key={item.code}>
-              <div
-                className={`${styles.customizeQueryCheckbox} ${styles.customizeQueryCheckbox} ${styles.hideCheckboxLabel}`}
-                onClick={(e) => {
-                  e.stopPropagation();
-                  toggleInclude(groupIndex, index);
-                }}
-              >
-                {item.include && (
-                  <Icon.Check
-                    className="usa-icon"
-                    style={{ backgroundColor: "white" }}
-                    size={4}
-                    color="#005EA2"
-                  />
-                )}
-              </div>
-              <div>{item.code}</div>
-              <div>{item.display}</div>
-            </div>
-          ))}
-        </div>
-      </div>
-    </div>
+    <Table className={`${styles.customizeQueryGridContainer}`}>
+      <thead className={`${styles.customizeQueryGridHeader} margin-top-10`}>
+        <th className={`${styles.accordionTableHeader}`}>Include</th>
+        <th className={`${styles.accordionTableHeader}`}>Code</th>
+        <th className={`${styles.accordionTableHeader}`}>Display</th>
+      </thead>
+      <tbody className="display-flex flex-column">
+        {group.items.map((item, index) => (
+          <tr className={`${styles.customizeQueryGridRow}`} key={item.code}>
+            <td
+              className={`${styles.customizeQueryCheckbox}`}
+              onClick={(e) => {
+                e.stopPropagation();
+                toggleInclude(groupIndex, index);
+              }}
+            >
+              {item.include && (
+                <Icon.Check
+                  className="usa-icon"
+                  style={{ backgroundColor: "white" }}
+                  size={4}
+                  color="#005EA2"
+                />
+              )}
+            </td>
+            <div>{item.code}</div>
+            <div>{item.display}</div>
+          </tr>
+        ))}
+      </tbody>
+    </Table>
   );
 };
